+++ lib/data/src/macro_def_or_undef.cpp
@@ -1,6 +1,3 @@
-#ifndef METASHELL_DATA_ENGINE_CONFIG_HPP
-#define METASHELL_DATA_ENGINE_CONFIG_HPP
-
 // Metashell - Interactive C++ template metaprogramming shell
 // Copyright (C) 2019, Abel Sinkovics (abel@sinkovics.hu)
 //
@@ -17,28 +14,28 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <metashell/data/command_line_argument_list.hpp>
-#include <metashell/data/include_config.hpp>
-#include <metashell/data/language_standard.hpp>
 #include <metashell/data/macro_def_or_undef.hpp>
-#include <metashell/data/standard_headers_allowed.hpp>
 
-#include <vector>
+#include <iostream>
 
 namespace metashell
 {
   namespace data
   {
-    struct engine_config
+    command_line_argument clang_argument(const macro_def_or_undef& m_)
     {
-      include_config includes;
+      return mpark::visit(
+          [](const auto& d_) { return clang_argument(d_); }, m_);
+    }
 
-      std::vector<macro_def_or_undef> macros;
-      language_standard standard = language_standard::cpp98;
-      standard_headers_allowed use_standard_headers =
-          standard_headers_allowed::all;
-    };
+    std::string to_string(const macro_def_or_undef& m_)
+    {
+      return mpark::visit([](const auto& d_) { return to_string(d_); }, m_);
+    }
+
+    std::ostream& operator<<(std::ostream& out_, const macro_def_or_undef& m_)
+    {
+      return out_ << to_string(m_);
+    }
   }
 }
-
-#endif